Created a function that takes the string references, converts them into i32 integers, multiplies them together following this formula to calculate the byte size of the cache:
2ˢ * 2ᵇ * E

This uses the .pow function which requires an unsigned integer, and as such I use the method try_into in order to attempt converting the i32 into an unsigned int, with the program panicking if I am unable to do so (these numbers should always be positive in any case).

I then print the result to the screen which in the following trace:

-s 4 -E 1 -b 4 -t traces/yi.trace

results in a 256 byte cache.